split java server
